requirements.txt
absl-py==0.9.0
altgraph==0.17
astroid==2.2.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
colorama==0.4.1
cssselect==1.1.0
cv==1.0.0
cx-Freeze==6.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
gast==0.3.3
get==2019.4.13
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.6.1
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.2
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
numpy==1.19.1
oauthlib==3.1.0
opencv-python==3.4.2.17
opt-einsum==3.2.1
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
post==2019.4.13
prompt-toolkit==3.0.5
protobuf==3.12.2
public==2019.4.13
py2exe==0.9.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
Pygments==2.6.1
pyinstaller==4.1.dev0
pyinstaller-hooks-contrib==2020.7
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.6
pyperclip==1.7.0
pypiwin32==223
pyquery==1.4.1
python-dateutil==2.8.1
pywin32==227
pywin32-ctypes==0.2.0
query-string==2019.4.13
redis==3.5.3
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.15
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
torch==1.5.0+cu92
torchvision==0.6.0+cu92
traitlets==4.3.3
typed-ast==1.4.0
urlfetch==1.1.3
urllib3==1.25.5
wcwidth==0.2.5
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
